refactor(inspections-api): extract core sub-router + finalize thin aggregator
Residual core routes (get/create/update/delete by id, property facts, results,
template snapshot, rating switch, recommendations, clone, wizard, patch item
field, preflight, results batch, /:id/full, presence/ws — 19 routes) move to
server/api/inspections/core.ts. inspections.ts is now a 38-line aggregator that
mounts the six sub-routers via .route('/', ...), preserving the exact external
path surface and the InspectionsApi RPC type. Handler bodies byte-identical.

Verified: route table identical (90 routes, method+path) vs the original; full
type-check, test:unit (1937), test:workers (41), and lint all green.
